+++ db3_5030_5152/db3_5030_5152/src/main/java/mye30/project/db3_5030_5152/service/JournalService.java
@@ -17,35 +17,35 @@ public interface JournalService {
 
     List<Object[]> findJournalByFirstYear(String journal_name);
 
-    int findNumOfJournalAuthors(String journal_name);
+    Integer findNumOfJournalAuthors(String journal_name);
 
-    int findAvgAuthorsByJournal(String journal_name);
+    Double findAvgAuthorsByJournal(String journal_name);
 
-    int findAvgAuthorsByYear(String journal_name);
+    Double findAvgAuthorsByYear(String journal_name);
 
     List<Article> findJournalArticles(String journal_name);
 
     List<Author> findJournalAuthors(String journal_name);
 
-    int findAvgJournalArticles(String journal_name);
+    Double findAvgJournalArticles(String journal_name);
 
-    int findNumOfJournalAuthorsRange(String journal_name, int y1, int y2);
+    Integer findNumOfJournalAuthorsRange(String journal_name, int y1, int y2);
 
-    int findAvgAuthorsByJournalRange(String journal_name, int y1, int y2);
+    Double findAvgAuthorsByJournalRange(String journal_name, int y1, int y2);
 
-    int findAvgAuthorsByYearRange(String journal_name, int y1, int y2);
+    Double findAvgAuthorsByYearRange(String journal_name, int y1, int y2);
 
     List<Article> findJournalArticlesRange(String journal_name, int y1, int y2);
 
     List<Author> findJournalAuthorsRange(String journal_name, int y1, int y2);
 
-    int findAvgJournalArticlesRange(String journal_name, int y1, int y2);
+    Double findAvgJournalArticlesRange(String journal_name, int y1, int y2);
 
     // TODO REPORT
 
     List<Object[]> findNumOfJournalByCategory(String categoryName);
 
-    int findPublisherPublications(String publisher);
+    Integer findPublisherPublications(String publisher);
 
     List<Object[]> findAvgAuthorsNumArticlesByYear(String journal_name);
 }
